+++ priv/repo/migrations/20150719222045_add_email_and_password_to_players.exs
@@ -1,0 +1,10 @@
+defmodule ApathyDrive.Repo.Migrations.AddEmailAndPasswordToPlayers do
+  use Ecto.Migration
+
+  def change do
+    alter table(:players) do
+      add :email, :text
+      add :password, :text
+    end
+  end
+end