+++ priv/repo/migrations/20150719231641_create_character.exs
@@ -1,0 +1,17 @@
+defmodule ApathyDrive.Repo.Migrations.CreateCharacter do
+  use Ecto.Migration
+
+  def change do
+    create table(:characters) do
+      add :name,       :text
+      add :player_id,  :integer
+      add :race_id,    :integer
+      add :class_id,   :integer
+      add :experience, :integer
+      add :alignment,  :integer
+
+      timestamps
+    end
+
+  end
+end