+++ priv/repo/migrations/20150726015637_create_race.exs
@@ -1,0 +1,14 @@
+defmodule ApathyDrive.Repo.Migrations.CreateRace do
+  use Ecto.Migration
+
+  def change do
+    create table(:races) do
+      add :name, :text
+      add :description, :text
+      add :properties, :jsonb
+
+      timestamps
+    end
+
+  end
+end