+++ priv/repo/migrations/20150801215343_remove_players_characters_races_classes.exs
@@ -1,0 +1,10 @@
+defmodule ApathyDrive.Repo.Migrations.RemovePlayersCharactersRacesClasses do
+  use Ecto.Migration
+
+  def change do
+    drop table(:players)
+    drop table(:characters)
+    drop table(:races)
+    drop table(:classes)
+  end
+end