+++ web/controllers/page_controller.ex
@@ -1,10 +1,6 @@
 defmodule ApathyDrive.PageController do
   use ApathyDrive.Web, :controller
   alias ApathyDrive.Repo
-  import Systems.Text
-
-  plug :scrub_params, "spirit" when action in [:update_spirit]
-  plug :action
 
   def index(conn, _params) do
     render conn, "index.html"
@@ -21,7 +17,7 @@ defmodule ApathyDrive.PageController do
           %Spirit{id: id, name: nil} ->
             conn
             |> put_session(:current_spirit, id)
-            |> redirect(to: "/create")
+            |> redirect(to: spirit_path(conn, :edit))
           %Spirit{} ->
             render conn, "game.html", []
           nil ->
@@ -31,53 +27,5 @@ defmodule ApathyDrive.PageController do
         end
     end
   end
-
-  def edit_spirit(conn, _params) do
-    spirit = Repo.get(Spirit, get_session(conn, :current_spirit))
-    changeset = Spirit.changeset(spirit)
-    render conn, "edit.html", changeset: changeset
-  end
-
-  def update_spirit(conn, %{"spirit" => spirit_params}) do
-    spirit = Repo.get(Spirit, get_session(conn, :current_spirit))
-
-    alignment = case spirit_params["faction"] do
-      "Demon" ->
-        "evil"
-      "Angel" ->
-        "good"
-      "Elemental" ->
-        "neutral"
-      _ ->
-        nil
-    end
-
-    spirit_params =
-      spirit_params
-      |> Map.put("name", capitalize_first(spirit_params["name"] || ""))
-      |> Map.put("alignment", alignment)
-
-    changeset = Spirit.changeset(spirit, spirit_params)
-
-    if changeset.valid? do
-      spirit =
-        changeset
-        |> Repo.update
-
-      case :global.whereis_name(:"spirit_#{spirit.id}") do
-        :undefined ->
-          conn
-          |> redirect(to: game_path(conn, :game))
-        pid ->
-          send(pid, {:reroll, name: spirit.name, faction: spirit.faction, alignment: spirit.alignment})
-
-          conn
-          |> redirect(to: game_path(conn, :game))
-      end
-    else
-      render conn, "edit.html", changeset: changeset
-    end
-  end
-
 
 end