+++ web/models/room.ex
@@ -1,9 +1,8 @@
 defmodule Room do
   require Logger
-  use Ecto.Model
+  use ApathyDrive.Web, :model
   use GenServer
   use Timex
-  alias ApathyDrive.Repo
   alias ApathyDrive.PubSub
 
   schema "rooms" do
@@ -37,7 +36,7 @@ defmodule Room do
 
     room = if room.lair_monsters do
       PubSub.subscribe(self, "rooms:lairs")
-      TimerManager.call_every(room, {:spawn_monsters, 60_000, fn -> send(self, {:spawn_monsters, Date.now |> Date.convert(:secs)}) end})
+      TimerManager.call_every(room, {:spawn_monsters, 60_000, fn -> send(self, {:spawn_monsters, Date.now |> Date.to_secs}) end})
     else
       room
     end
@@ -347,7 +346,7 @@ defmodule Room do
     room = room
            |> Map.put(:lair_next_spawn_at, Date.now
                                            |> Date.shift(mins: room.lair_frequency)
-                                           |> Date.convert(:secs))
+                                           |> Date.to_secs)
 
     {:noreply, room}
   end
@@ -365,7 +364,7 @@ defmodule Room do
          end)
 
       if room.lair_monsters do
-        send(this, {:spawn_monsters, Date.now |> Date.convert(:secs)})
+        send(this, {:spawn_monsters, Date.now |> Date.to_secs})
       end
 
     end
@@ -534,7 +533,7 @@ defmodule Room do
       room =
         room
         |> Map.put(:lair_faction, faction)
-        |> Repo.update
+        |> Repo.update!
 
       send(monster, {:scroll, "<p>You capture the lair for your faction!</p>"})
       ApathyDrive.Factions.update_war_status