+++ web/models/spirit.ex
@@ -1,11 +1,10 @@
 defmodule Spirit do
-  use Ecto.Model
   use GenServer
   use ApathyDrive.Web, :model
 
   require Logger
   import Systems.Text
-  alias ApathyDrive.Repo
+  import Comeonin.Bcrypt
   alias ApathyDrive.PubSub
 
   @idle_threshold 60
@@ -14,6 +13,8 @@ defmodule Spirit do
     belongs_to :room, Room
     field :name,              :string
     field :alignment,         :string
+    field :email,             :string
+    field :password,          :string
     field :external_id,       :string
     field :experience,        :integer, default: 0
     field :level,             :integer, default: 1
@@ -60,19 +61,41 @@ defmodule Spirit do
     |> validate_length(:name, min: 1, max: 18)
   end
 
+  def sign_up_changeset(model, params \\ :empty) do
+    model
+    |> cast(params, ~w(email password), [])
+    |> validate_format(:email, ~r/@/)
+    |> validate_length(:email, min: 3, max: 100)
+    |> validate_length(:password, min: 6)
+    |> validate_confirmation(:password)
+  end
+
+  def sign_in(email, password) do
+    player = Repo.get_by(Player, email: email)
+    sign_in?(player, password) && player
+  end
+
+  def sign_in?(%Spirit{password: stored_hash}, password) do
+    checkpw(password, stored_hash)
+  end
+
+  def sign_in?(nil, _password) do
+    dummy_checkpw
+  end
+
   def find_or_create_by_external_id(external_id) do
     case Repo.one from s in Spirit, where: s.external_id == ^external_id do
       %Spirit{} = spirit ->
         spirit
       nil ->
         %Spirit{room_id: Room.start_room_id, external_id: external_id}
-        |> Repo.insert
+        |> Repo.insert!
     end
   end
 
   def save(spirit) when is_pid(spirit), do: spirit |> value |> save
   def save(%Spirit{id: id} = spirit) when is_integer(id) do
-    Repo.update(spirit)
+    Repo.update!(spirit)
   end
   def save(%Spirit{} = spirit), do: spirit