+++ web/web.ex
@@ -2,26 +2,28 @@ defmodule ApathyDrive.Web do
   @moduledoc """
   A module that keeps using definitions for controllers,
   views and so on.
-
   This can be used in your application as:
-
-      use MyApp.Web, :controller
-      use MyApp.Web, :view
-
-  Keep the definitions in this module short and clean,
-  mostly focused on imports, uses and aliases.
+      use <%= application_module %>.Web, :controller
+      use <%= application_module %>.Web, :view
+  The definitions below will be executed for every view,
+  controller, etc, so keep them short and clean, focused
+  on imports, uses and aliases.
+  Do NOT define functions inside the quoted expressions
+  below.
   """
 
   def view do
     quote do
       use Phoenix.View, root: "web/templates"
 
+      # Import convenience functions from controllers
+      import Phoenix.Controller, only: [get_csrf_token: 0, get_flash: 2, view_module: 1]
+
       # Import URL helpers from the router
       import ApathyDrive.Router.Helpers
 
-      # Import all HTML functions (forms, tags, etc)
+      # Use all HTML functionality (forms, tags, etc)
       use Phoenix.HTML
-      import Phoenix.Controller, only: [get_flash: 2]
 
       def list_input(form, field, opts) when is_atom(field) and is_list(opts) do
         opts =
@@ -76,13 +78,35 @@ defmodule ApathyDrive.Web do
     quote do
       use Phoenix.Controller
 
-      # Import URL helpers from the router
+      # Alias the data repository and import query/model functions
+      alias ApathyDrive.Repo
+      import Ecto.Model
+      import Ecto.Query, only: [from: 2]
       import ApathyDrive.Router.Helpers
     end
   end
 
   def model do
     quote do
+      use Ecto.Model
+      use Timex.Ecto.Timestamps
+      alias ApathyDrive.Repo
+    end
+  end
+
+  def router do
+    quote do
+      use Phoenix.Router
+    end
+  end
+
+  def channel do
+    quote do
+      use Phoenix.Channel
+      # Alias the data repository and import query/model functions
+      alias ApathyDrive.Repo
+      import Ecto.Model
+      import Ecto.Query, only: [from: 2]
     end
   end